00 Quickstart/requirements.txt
asgiref==3.5.2
autopep8==1.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
commonmark==0.9.1
defusedxml==0.7.1
Django==4.0.4
djangorestframework==3.13.1
httpie==3.2.1
idna==3.3
multidict==6.0.2
pycodestyle==2.8.0
Pygments==2.12.0
PySocks==1.7.1
pytz==2022.1
requests==2.27.1
requests-toolbelt==0.9.1
rich==12.4.4
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.9
